replaced u32-types with i32

+++ src/cpu/x64/param.rs
@@ -13,11 +13,11 @@ pub fn next_param_offset(param_offset: i32, _: BuiltinType) -> i32 {
 }
 
 // needed memory on stack for function call in caller
-pub fn reserve_stack_for_call(args: &[Box<Expr>]) -> u32 {
+pub fn reserve_stack_for_call(args: &[Box<Expr>]) -> i32 {
     let params_on_stack = args.len() as i32 - cpu::REG_PARAMS.len() as i32;
 
     if params_on_stack > 0 {
-        (params_on_stack as u32) * 8
+        params_on_stack * 8
     } else {
         0
     }
